work around circular dependencies
