+++ packages/services/api/src/modules/schema/providers/schema-publisher.ts
@@ -2,10 +2,11 @@ import { Inject, Injectable, Scope } from 'graphql-modules';
 import lodash from 'lodash';
 import promClient from 'prom-client';
 import { Change, CriticalityLevel } from '@graphql-inspector/core';
+import { SchemaCheck } from '@hive/storage';
 import * as Sentry from '@sentry/node';
 import type { Span } from '@sentry/types';
 import * as Types from '../../../__generated__/types';
-import { Project, ProjectType, Schema, SchemaCheck, Target } from '../../../shared/entities';
+import { Project, ProjectType, Schema, Target } from '../../../shared/entities';
 import { HiveError } from '../../../shared/errors';
 import { bolderize } from '../../../shared/markdown';
 import { sentry } from '../../../shared/sentry';
@@ -21,7 +22,7 @@ import { Logger } from '../../shared/providers/logger';
 import { Mutex } from '../../shared/providers/mutex';
 import { Storage, type TargetSelector } from '../../shared/providers/storage';
 import { TargetManager } from '../../target/providers/target-manager';
-import { toGraphQLSchemaCheck } from '../resolvers';
+import { toGraphQLSchemaCheck } from '../to-graphql-schema-check';
 import { ArtifactStorageWriter } from './artifact-storage-writer';
 import type { SchemaModuleConfig } from './config';
 import { SCHEMA_MODULE_CONFIG } from './config';
