+++ packages/services/api/src/modules/shared/providers/storage.ts
@@ -1,6 +1,7 @@
 import { Injectable } from 'graphql-modules';
 import { Change } from '@graphql-inspector/core';
 import type { PolicyConfigurationObject } from '@hive/policy';
+import type { SchemaCheck, SchemaCheckInput, SchemaCompositionError } from '@hive/storage';
 import type {
   AddAlertChannelInput,
   AddAlertInput,
@@ -24,9 +25,6 @@ import type {
   PersistedOperation,
   Project,
   Schema,
-  SchemaCheck,
-  SchemaCheckInput,
-  SchemaCompositionError,
   SchemaLog,
   SchemaPolicy,
   SchemaVersion,
